Extract TF/PT common example code
