+++ examples/common/README.md
@@ -1,0 +1,2 @@
+This directory contains common code for example scripts.
+See [other directories at the same level](./..) for actual example scripts that you can launch to evaluate NNCF for various backend frameworks and use cases.
