+++ examples/tensorflow/object_detection/models/retinanet_config.py
@@ -10,9 +10,7 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 """
-
-from examples.tensorflow.common.sample_config import SampleConfig
-
+from examples.common.sample_config import SampleConfig
 
 RETINANET_CONFIG = SampleConfig({
     "preprocessing": {
