+++ examples/tensorflow/object_detection/models/yolo_v4_config.py
@@ -11,7 +11,7 @@
  limitations under the License.
 """
 
-from examples.tensorflow.common.sample_config import SampleConfig
+from examples.common.sample_config import SampleConfig
 
 YOLOV4_CONFIG = SampleConfig({
     "preprocessing": {
