+++ tests/torch/quantization/test_hawq_precision_init.py
@@ -32,7 +32,7 @@ from torchvision.models import resnet50
 from torchvision.transforms import transforms
 
 from examples.torch.classification.main import create_cifar
-from examples.torch.common.sample_config import SampleConfig
+from examples.common.sample_config import SampleConfig
 from examples.torch.object_detection.models.ssd_vgg import SSD_VGG
 from nncf.torch import register_default_init_args
 from nncf.torch.checkpoint_loading import load_state
