+++ tests/torch/test_models/ssd_mobilenet.py
@@ -13,7 +13,7 @@
 
 import torch
 from torch import nn
-from examples.torch.common.sample_config import SampleConfig
+from examples.common.sample_config import SampleConfig
 
 from examples.torch.object_detection.layers.modules.ssd_head import MultiOutputSequential, SSDDetectionOutput
 from nncf.torch.checkpoint_loading import load_state
