+++ tests/torch/test_models/ssd_vgg.py
@@ -14,7 +14,7 @@
 import os
 import torch
 from torch import nn
-from examples.torch.common.sample_config import SampleConfig
+from examples.common.sample_config import SampleConfig
 
 from examples.torch.object_detection.layers import L2Norm
 from examples.torch.object_detection.layers.modules.ssd_head import MultiOutputSequential, SSDDetectionOutput
